rev-sales-pitch-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
elasticsearch
langchain
langchain-community
langchain-openai
python-multipart